// public/models/CatalogThemeColors.ts
/**
 * export the class catalog theme colors
 */
export class CatalogThemeColors {
    PrimaryColor1: string;

    PrimaryColor2: string;

    PrimaryColor3: string;

    AccentColor1: string;

    AccentColor2: string;

    AccentColor3: string;

    TextColor: string;

    SecondaryTextColor: string;

    AlternateTextColor: string;

    CanvasColor: string;

    AlternateCanvasColor: string;

    AlternateColor1: string;

    AlternateColor2: string;

    BorderColor: string;

    DisabledColor: string;
}